+++ src/hooks/useDropDown.ts
@@ -1,0 +1,21 @@
+import { useEffect, useRef, useState } from 'react';
+
+export const useDropdown = () => {
+  const [showDropdown, setShowDropdown] = useState(false);
+  const dropDownRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (dropDownRef.current && !dropDownRef.current.contains(event.target as Node)) {
+        setShowDropdown(false);
+      }
+    };
+
+    document.addEventListener('click', handleClickOutside);
+    return () => {
+      document.removeEventListener('click', handleClickOutside);
+    };
+  }, []);
+
+  return [showDropdown, setShowDropdown, dropDownRef] as const;
+};
